+++ CHANGES.rst
@@ -1,6 +1,7 @@
 0.1.1 (UNRELEASED)
 ------------------
 
+- Ignore InvalidSessionIdException while quitting webdriver
 - Do not use deprecated assert method
 - Support version 3
 - Use geckodriver to test firefox